Move all code related to specific achievements out of the view and some of it out of data as well
